chore: bump flake inputs, cargo deps, and infra pins
flake inputs:
- nixpkgs 6201e203→b12141ef (apr 1→18)
- nix 2.34.2→2.34.6
- rust-overlay, nixhelm, advisory-db, tracey-src, lix,
  nix-unstable, git-hooks-nix, treefmt-nix

cargo:
- drop reqwest rustls-native-certs feature (removed in 0.13.2)
- drop legacy rustls feature on aws-sdk-*/aws-config; rustls 0.21
  + rustls-webpki 0.101.7 removed from tree (clears
  RUSTSEC-2026-0098/0099/0104)
- floor aws-sdk-s3 at 1.130 to stop resolver downgrading via crc 3.4
- rand 0.8.6/0.9.4/0.10.1, rustls-webpki 0.103.13 (RUSTSEC-2026-0097)
- fastcdc 4.0: CHUNK_{MIN,AVG,MAX} retyped u32→usize
- deny.toml: drop stale aws-lc-sys ignores
- regen workspace-hack, Cargo.json, fuzz lockfiles

infra pins (pins.nix + tfvars + tf):
- cilium 1.19.2→1.19.3 (+ docker-pulled digests)
- gateway-api v1.2.1→v1.4.0; cilium-render now derives from
  pins.nix (was hardcoded → drifted vs EKS)
- karpenter 1.10.0→1.11.1, aws-lbc 3.1.0→3.2.2
- nodeadm v20260318→v20260415
- kube-prometheus-stack 75.6.0→83.7.0 (+ crds.upgradeJob.enabled)
- external-secrets 2.2.0→2.3.0
- aurora-postgresql 16.4→17.9 (+ allow_major_version_upgrade)
- bitnami subchart 18.5.6→18.6.1
- tracey version date 2026-03-13→04-11

+++ rio-store/src/chunker.rs
@@ -29,14 +29,14 @@
 use fastcdc::v2020::FastCDC;
 
 /// Minimum chunk size. FastCDC won't cut before this many bytes.
-pub const CHUNK_MIN: u32 = 16 * 1024;
+pub const CHUNK_MIN: usize = 16 * 1024;
 
 /// Average chunk size. The "expected" boundary spacing.
-pub const CHUNK_AVG: u32 = 64 * 1024;
+pub const CHUNK_AVG: usize = 64 * 1024;
 
 /// Maximum chunk size. Hard cap — FastCDC forces a cut at this length
 /// even if no content-defined boundary is found.
-pub const CHUNK_MAX: u32 = 256 * 1024;
+pub const CHUNK_MAX: usize = 256 * 1024;
 
 /// A single chunk: its BLAKE3 hash and a borrowed slice into the source.
 ///
@@ -147,14 +147,14 @@ mod tests {
             // nothing to enforce it (can't cut past EOF).
             if i < chunks.len() - 1 {
                 assert!(
-                    chunk.data.len() >= CHUNK_MIN as usize,
+                    chunk.data.len() >= CHUNK_MIN,
                     "chunk {i} is {} bytes, below CHUNK_MIN {}",
                     chunk.data.len(),
                     CHUNK_MIN
                 );
             }
             assert!(
-                chunk.data.len() <= CHUNK_MAX as usize,
+                chunk.data.len() <= CHUNK_MAX,
                 "chunk {i} is {} bytes, above CHUNK_MAX {}",
                 chunk.data.len(),
                 CHUNK_MAX
